requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
python-dotenv
requests
